Better error for `:named` attributes in binders
Currently, when a `:named` attribute is used in a binder, the parser
complains about an illegal argument. This is because an argument check
in the `SymbolManager` fails. This is not very user friendly. This
commit makes the error message clearer for the user:

> Cannot name a term in a binder (e.g., quantifiers, definitions)

To do this, the commit changes the return type for
`SymbolManager::setExpressionName` to include more information that can
then be used by the parser to generate an appropriate error message.